+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/jaxb/impl/projection/impl/ExpectedExpressionProjection.java
@@ -15,23 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.jaxb.impl.projection;
+package org.apache.shardingsphere.sql.parser.integrate.jaxb.impl.projection.impl;
 
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.generic.AbstractExpectedSegment;
+import org.apache.shardingsphere.sql.parser.integrate.jaxb.impl.projection.ExpectedProjection;
 
 import javax.xml.bind.annotation.XmlAttribute;
 
 @Getter
 @Setter
-public class ExpectedAggregationProjection extends AbstractExpectedSegment implements ExpectedProjection {
-    
-    @XmlAttribute
-    private String type;
-    
-    @XmlAttribute(name = "inner-expression-start-index")
-    private int innerExpressionStartIndex;
+public final class ExpectedExpressionProjection extends AbstractExpectedSegment implements ExpectedProjection {
     
     @XmlAttribute
     private String alias;
