+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/jaxb/impl/projection/impl/ExpectedTopProjection.java
@@ -15,25 +15,31 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.jaxb.impl.projection;
+package org.apache.shardingsphere.sql.parser.integrate.jaxb.impl.projection.impl;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.apache.shardingsphere.sql.parser.integrate.jaxb.generic.AbstractExpectedDelimiterSegment;
+import org.apache.shardingsphere.sql.parser.integrate.jaxb.generic.AbstractExpectedSegment;
+import org.apache.shardingsphere.sql.parser.integrate.jaxb.impl.projection.ExpectedProjection;
 
 import javax.xml.bind.annotation.XmlAttribute;
-import javax.xml.bind.annotation.XmlElement;
 
 @Getter
 @Setter
-public final class ExpectedColumnProjection extends AbstractExpectedDelimiterSegment implements ExpectedProjection {
+public final class ExpectedTopProjection extends AbstractExpectedSegment implements ExpectedProjection {
     
-    @XmlAttribute
-    private String name;
+    @XmlAttribute(name = "top-value")
+    private Long topValue;
     
-    @XmlAttribute
+    @XmlAttribute(name = "top-parameter-index")
+    private Integer topParameterIndex;
+    
+    @XmlAttribute(name = "top-value-start-index")
+    private int topValueStartIndex;
+    
+    @XmlAttribute(name = "top-value-stop-index")
+    private int topValueStopIndex;
+    
+    @XmlAttribute(name = "alias")
     private String alias;
-    
-    @XmlElement(name = "owner")
-    private ExpectedTableSegment owner;
 }
